+++ docs-src/.vitepress/src/config/en.ts
@@ -130,6 +130,7 @@ export const enConfig = defineConfig({
                     { text: 'Cell Edit', link: '/cell-edit' },
                     { text: 'Panel Tree', link: '/panel-tree' },
                     { text: 'Lazy Load', link: '/lazy-load' },
+                    { text: 'Realtime Merge Cells(✨NEW)', link: '/realtime-merge-cells' },
                 ]
             }
         },