+++ docs-src/.vitepress/src/config/ko.ts
@@ -130,6 +130,7 @@ export const koConfig = defineConfig({
                     { text: '셀 편집', link: '/cell-edit' },
                     { text: '패널 트리', link: '/panel-tree' },
                     { text: '지연 로딩', link: '/lazy-load' },
+                    { text: '실시간 셀 병합(✨NEW)', link: '/realtime-merge-cells' },
                 ]
             }
         },